+++ src/pyiem/plot/geoplot.py
@@ -856,6 +856,7 @@ class MapPlot:
           vals (list): Data values for the points to use for colormapping
           clevs (list): Levels to use for ramp
           **kwargs: additional options
+            draw_colorbar (bool, optional): Draw a colorbar, default True.
         """
         cmap = stretch_cmap(
             kwargs.get("cmap"), clevs, extend=kwargs.get("extend")
@@ -872,10 +873,15 @@ class MapPlot:
             zorder=Z_OVERLAY,
         )
         kwargs.pop("cmap", None)
-        self.draw_colorbar(clevs, cmap, norm, **kwargs)
+        if kwargs.pop("draw_colorbar", True):
+            self.draw_colorbar(clevs, cmap, norm, **kwargs)
 
     def hexbin(self, lons, lats, vals, clevs, **kwargs):
-        """hexbin wrapper"""
+        """hexbin wrapper.
+
+        Args:
+            draw_colorbar (bool,optional): Draw colorbar default True.
+        """
         cmap = kwargs.pop("cmap", None)
         cmap = stretch_cmap(cmap, clevs, extend=kwargs.get("extend"))
         norm = mpcolors.BoundaryNorm(clevs, cmap.N)
@@ -888,7 +894,8 @@ class MapPlot:
             cmap=cmap,
             zorder=Z_FILL,
         )
-        self.draw_colorbar(clevs, cmap, norm, **kwargs)
+        if kwargs.pop("draw_colorbar", True):
+            self.draw_colorbar(clevs, cmap, norm, **kwargs)
         return _hex
 
     def pcolormesh(self, lons, lats, vals, clevs, **kwargs):
@@ -898,6 +905,9 @@ class MapPlot:
         tack on an extra row and column to make matplotlib happy. If you do
         not want this, then pass your own lons + lats that is 1 column and 1
         row greater than vals.
+
+        Args:
+            draw_colorbar (bool,optional): Draw colorbar default True.
         """
         cmap = stretch_cmap(
             kwargs.get("cmap"), clevs, extend=kwargs.get("extend")
@@ -927,7 +937,8 @@ class MapPlot:
         if kwargs.get("clip_on", True):
             self.draw_mask()
         kwargs.pop("cmap", None)
-        self.draw_colorbar(clevs, cmap, norm, **kwargs)
+        if kwargs.pop("draw_colorbar", True):
+            self.draw_colorbar(clevs, cmap, norm, **kwargs)
         return res
 
     def draw_mask(self, sector=None):
@@ -981,6 +992,7 @@ class MapPlot:
           ilabel (boolean,optional): Should we label contours
           iline (boolean,optional): should we draw contour lines
           lblformat (str,optional): Format string for labeling contours, %.0f.
+            draw_colorbar (bool,optional): Draw colorbar default True.
 
         Returns:
           vals (np.array): The values used for plotting, maybe after gridding
@@ -1046,7 +1058,8 @@ class MapPlot:
         if kwargs.get("clip_on", True):
             self.draw_mask()
         kwargs.pop("cmap", None)
-        self.draw_colorbar(clevs, cmap, norm, **kwargs)
+        if kwargs.pop("draw_colorbar", True):
+            self.draw_colorbar(clevs, cmap, norm, **kwargs)
         return vals
 
     def fill_climdiv(self, data, **kwargs):
@@ -1070,8 +1083,8 @@ class MapPlot:
           color(dict, optional): Hard code what each UGC should display as
             for color.
           is_firewx(bool, optional): Are we plotting fire weather zones?
-          nocbar (bool, optional): Should a color bar be generated, default is
-            `True`.
+          draw_colorbar (bool, optional): Should a color bar be generated,
+            default is `True`.
           plotmissing(bool, optional): Should missing UGC data be plotted?
           labels(dict, optional): UGC indexed dictionary to use for labeling.
           lblformat(str, optional): Format string for labels, default %s.