+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 All notable changes to this library are documented in this file.
 
+## Unreleased Version
+
+### API Changes
+
+- For `pyiem.plot.geoplot.fill_ugcs` changed `nocbar` kwarg to `draw_colorbar`.
+
 ## **1.9.0** (13 Dec 2021)
 
 ### API Changes